# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FrankensteinDB API",
    "version": "0.2.0",
    "description": "Typed JSON API. SQLite is authoritative for writes; every supported read executes through Tantivy."
  },
  "servers": [{ "url": "http://127.0.0.1:8080" }],
  "security": [{ "bearerAuth": [] }, {}],
  "paths": {
...
    "/api/v1/tables": {
      "get": { "summary": "List published table definitions", "responses": { "200": { "description": "Tables" } } },
      "post": { "summary": "Create table", "requestBody": { "$ref": "#/components/requestBodies/Table" }, "responses": { "201": { "description": "Created" } } }
    },
    "/api/v1/tables/{table}": {
      "parameters": [{ "$ref": "#/components/parameters/Table" }],
      "get": { "summary": "Get table definition", "responses": { "200": { "description": "Table" } } },
      "delete": { "summary": "Drop table", "responses": { "204": { "description": "Dropped" } } }
    },
...
    "/api/v1/jobs/{id}": { "parameters": [{ "$ref": "#/components/parameters/JobId" }], "get": { "summary": "Get job", "responses": { "200": { "$ref": "#/components/responses/Job" } } }, "delete": { "summary": "Cancel job", "responses": { "204": { "description": "Cancelled" } } } },
...
    "/api/v1/jobs/{id}/artifact": { "parameters": [{ "$ref": "#/components/parameters/JobId" }], "get": { "summary": "Download completed job artifact", "responses": { "200": { "description": "Artifact stream" } } } },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
    "parameters": {
      "Table": { "name": "table", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "JobId": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
...
    },
    "requestBodies": {
...
      "Table": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TableDef" } } } }
    },
    "responses": {
...
      "Job": { "description": "Persistent job", "content": { "application/json": { "schema": { "type": "object" } } } },
...
    },
    "schemas": {
      "TableDef": { "type": "object", "required": ["name", "columns"], "properties": { "name": { "type": "string" }, "aliases": { "type": "array", "items": { "type": "string" }, "default": [] }, "document_store": { "$ref": "#/components/schemas/DocumentStore" }, "columns": { "type": "array", "items": { "$ref": "#/components/schemas/ColumnDef" } } } },
      "DocumentStore": { "type": "object", "properties": { "compression": { "enum": ["lz4", "zstd", "none"], "default": "lz4" }, "zstd_level": { "type": ["integer", "null"], "description": "Zstd default level 3 when omitted" }, "block_size": { "type": "integer", "minimum": 1024, "maximum": 16777216, "default": 16384 }, "dedicated_thread": { "type": "boolean", "default": true } } },
      "ColumnDef": { "type": "object", "required": ["name", "data_type", "primary_key", "nullable"], "properties": { "name": { "type": "string" }, "data_type": { "enum": ["Integer", "Unsigned", "Real", "Text", "Boolean", "Date", "DateTime", "Timestamp", "TextArray", "IntegerArray", "UnsignedArray", "RealArray", "BooleanArray", "DateArray", "DateTimeArray", "TimestampArray", "BlobArray", "IpArray", "JsonArray", "FacetArray", "Blob", "Ip", "Json", "Facet"] }, "primary_key": { "type": "boolean" }, "nullable": { "type": "boolean" }, "analyzer": {}, "compact_raw": { "type": "boolean", "default": false }, "index": { "$ref": "#/components/schemas/IndexProfile" } } },
      "IndexProfile": { "type": "object", "properties": { "indexed": { "type": "boolean", "default": true }, "record": { "enum": ["basic", "frequencies", "positions"], "default": "positions" }, "stored": { "type": "boolean", "default": false } } },
...
    }
  }
}